gpu-service/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# HTTP client
httpx==0.27.2

# AI/ML
torch==2.1.0
transformers==4.36.0
accelerate==0.25.0

# Google Cloud
google-cloud-storage==2.10.0

# Utilities
pydantic==2.9.2
python-dotenv==1.0.1
